src/app/editor/editor.component.spec.ts
import { async, ComponentFixture, TestBed } from '@angular/core/testing';

import { EditorComponent } from './editor.component';

describe('EditorComponent', () => {
  let component: EditorComponent;
  let fixture: ComponentFixture<EditorComponent>;

  beforeEach(async(() => {
    TestBed.configureTestingModule({
      declarations: [EditorComponent]
    }).compileComponents();
  }));

  beforeEach(() => {
    fixture = TestBed.createComponent(EditorComponent);
    component = fixture.componentInstance;
    fixture.detectChanges();
  });

  it('should exist', () => {
    expect(component).toBeTruthy();
  });

  // Test properties existence
  it('Property translationsFolder should exist', () => {
    expect(component).toHaveProperty('translationsFolder');
  });
  it('Property translationFiles should exist', () => {
    expect(component).toHaveProperty('translationFiles', undefined);
  });
});
